# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SME Tax & Finance Assistant Algeria API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Company ─────────────────────────────────────────────────────────────────
  /company:
    get:
      operationId: getCompany
      tags: [company]
      summary: Get company profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCompany
      tags: [company]
      summary: Create or update company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /reminders/{id}/acknowledge:
    patch:
      operationId: acknowledgeReminder
      tags: [reminders]
      summary: Acknowledge a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"

  # ─── Knowledge Base ──────────────────────────────────────────────────────────
  /knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [knowledge]
      summary: List knowledge base items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: regime
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    get:
      operationId: getKnowledgeItem
      tags: [knowledge]
      summary: Get a knowledge base item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── News ────────────────────────────────────────────────────────────────────
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news and announcements
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update a news item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ─── Tax Rules ───────────────────────────────────────────────────────────────
...
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user role or status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isActive, createdAt]

    # ─── Company ───────────────────────────────────────────────────────────────
    UpdateCompanyBody:
      type: object
      properties:
        companyName:
          type: string
        activityType:
          type: string
          enum: [commerce, services, production, auto_entrepreneur]
        annualRevenue:
          type: ["number", "null"]
        employeeCount:
          type: ["integer", "null"]
        taxRegime:
          type: string
          enum: [real, simplified_real, forfaitaire]
        hasCnas:
          type: boolean
      required: [companyName, activityType, taxRegime]
    Company:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyName:
          type: string
        activityType:
          type: string
        annualRevenue:
          type: ["number", "null"]
        employeeCount:
          type: ["integer", "null"]
        taxRegime:
          type: string
        hasCnas:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, companyName, activityType, taxRegime, hasCnas, createdAt, updatedAt]

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        declarationType:
          type: string
          description: "G50, G12, G12BIS, CNAS, IRG, IBS, G51, G4"
        regime:
          type: string
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        dueDate:
          type: string
        daysUntilDue:
          type: integer
        status:
          type: string
          enum: [upcoming, due_today, overdue, completed]
        urgencyLevel:
          type: string
          enum: [normal, warning, critical, overdue]
      required: [id, declarationType, regime, title, titleAr, dueDate, daysUntilDue, status, urgencyLevel]

    # ─── Reminders ─────────────────────────────────────────────────────────────
    Reminder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        calendarEventId:
          type: integer
        reminderType:
          type: string
          enum: [seven_days, three_days, one_day, due_day, overdue]
        message:
          type: string
        messageAr:
          type: string
        isAcknowledged:
          type: boolean
        scheduledFor:
          type: string
        createdAt:
          type: string
        event:
          $ref: "#/components/schemas/CalendarEvent"
      required: [id, userId, calendarEventId, reminderType, message, messageAr, isAcknowledged, scheduledFor, createdAt, event]

    # ─── Knowledge Base ────────────────────────────────────────────────────────
    KnowledgeItem:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        questionAr:
          type: string
        answer:
          type: string
        answerAr:
          type: string
        tip:
          type: ["string", "null"]
        tipAr:
          type: ["string", "null"]
        category:
          type: string
        regime:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, question, questionAr, answer, answerAr, category, createdAt]

    # ─── News ──────────────────────────────────────────────────────────────────
    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        publishedAt:
          type: string
        createdAt:
          type: string
      required: [id, title, titleAr, content, contentAr, category, publishedAt, createdAt]
    CreateNewsBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        publishedAt:
          type: string
      required: [title, titleAr, content, contentAr, category]

    # ─── Tax Rules ─────────────────────────────────────────────────────────────
...
    UpdateAdminUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
...